Time: 0 ms (100.00%) | Memory: 19.8 MB (12.31%) - LeetSync

+++ 662-maximum-width-of-binary-tree/maximum-width-of-binary-tree.cpp
@@ -16,7 +16,7 @@ public:
         if (!root)
             return 0;
         int maxWidth = 0;
-        queue<pair<TreeNode*, int>> q;
+        queue<pair<TreeNode*, long long>> q;
 
         q.push({root, 0});
         while (!q.empty()) {
